Separated crud calls from server.js into routes folder and integrated routes into server.js using app.use #72
